requirements.txt
aiogram[fast]
python-dotenv
aiosqlite==0.21.0
